+++ src/app/app.module.ts
@@ -38,7 +38,7 @@ import {
   ServicesApiService,
   SupportService,
   SystemLogService,
-  UserService,
+  UserService
 } from './services';
 import { HttpsRequestInterceptor } from './services/http.request.interceptor';
 import { SharedService } from './services/shared.service';
@@ -55,8 +55,8 @@ export function pingServiceFactory(ping: PingService, sharedService: SharedServi
       sharedService.isServiceUp.next(true);
     })
     .catch(error => {
-      console.log('error: ', error);
-      if (error.status === 403) {
+      // Set isService to true, if response status code is non zero and not undefined
+      if (error && error.status && error.status !== 0) {
         sharedService.isServiceUp.next(true);
       } else {
         sharedService.isServiceUp.next(false);