+++ src/app/components/core/filter/add-filter-wizard/add-filter-wizard.component.ts
@@ -5,6 +5,9 @@ import { assign, reduce, sortBy, isEmpty } from 'lodash';
 import { AlertService, ConfigurationService, FilterService, ServicesApiService, ProgressBarService } from '../../../../services';
 import { ViewConfigItemComponent } from '../../configuration-manager/view-config-item/view-config-item.component';
 import { ValidateFormService } from '../../../../services/validate-form.service';
+import { concatMap, delayWhen, retryWhen, take, tap } from 'rxjs/operators';
+import { of, Subscription, throwError, timer } from 'rxjs';
+import { DocService } from '../../../../services/doc.service';
 
 @Component({
   selector: 'app-add-filter-wizard',
@@ -22,7 +25,7 @@ export class AddFilterWizardComponent implements OnInit {
   public isValidName = true;
   public payload: any;
   public selectedPluginDescription = '';
-  public plugin:any;
+  public plugin: any;
   public pluginData = [];
   public filesToUpload = [];
   public requestInProgress = false;
@@ -30,6 +33,12 @@ export class AddFilterWizardComponent implements OnInit {
   public stopLoading = false;
   public placeholderText = 'fetching available plugins...';
   public disabledBtn = false;
+
+  increment = 1;
+  maxRetry = 15;
+  initialDelay = 1000;
+
+  installPluginSub: Subscription;
 
   serviceForm = new FormGroup({
     name: new FormControl(),
@@ -48,6 +57,7 @@ export class AddFilterWizardComponent implements OnInit {
     private configService: ConfigurationService,
     private alertService: AlertService,
     private service: ServicesApiService,
+    private docService: DocService,
     private validateFormService: ValidateFormService,
     private ngProgress: ProgressBarService) { }
 
@@ -55,8 +65,8 @@ export class AddFilterWizardComponent implements OnInit {
     this.getCategories();
     this.serviceForm = this.formBuilder.group({
       name: ['', Validators.required],
-      plugin: [{value: '', disabled: false}, Validators.required],
-      pluginToInstall: [{value: null, disabled: false}, Validators.required]
+      plugin: [{ value: '', disabled: false }, Validators.required],
+      pluginToInstall: [{ value: null, disabled: false }, Validators.required]
     });
     this.getInstalledFilterPlugins();
   }
@@ -276,13 +286,7 @@ export class AddFilterWizardComponent implements OnInit {
       subscribe(
         (data: any) => {
           /** request done */
-          this.ngProgress.done();
-          this.alertService.closeMessage();
-          this.alertService.success(data.message, true);
-          this.getInstalledFilterPlugins(true);
-          this.serviceForm.controls.pluginToInstall.enable();
-          this.serviceForm.controls.plugin.enable();
-          this.disabledBtn = false;
+          this.monitorFilterPluginInstallationStatus(data, pluginName);
         },
         error => {
           /** request done */
@@ -296,6 +300,58 @@ export class AddFilterWizardComponent implements OnInit {
             this.alertService.error(error.statusText);
           }
         });
+  }
+
+  monitorFilterPluginInstallationStatus(data: any, pluginName: string) {
+    this.installPluginSub = this.service.monitorPluginInstallationStatus(data.statusLink)
+      .pipe(
+        take(1),
+        // checking the response object for plugin.
+        // if pacakge.status === 'in-progress' then
+        // throw an error to re-fetch:
+        tap((response: any) => {
+          if (response.packageStatus[0].status === 'in-progress') {
+            this.increment++;
+            throw response;
+          }
+        }),
+        retryWhen(result =>
+          result.pipe(
+            // only if a server returned an error, stop trying and pass the error down
+            tap(installStatus => {
+              if (installStatus.error) {
+                this.ngProgress.done();
+                this.alertService.closeMessage();
+                throw installStatus.error;
+              }
+            }),
+            delayWhen(() => {
+              const delay = this.increment * this.initialDelay;     // incremental
+              console.log(new Date().toLocaleString(), `retrying after ${delay} msec...`);
+              return timer(delay);
+            }), // delay between api calls
+            // Set the number of attempts.
+            take(this.maxRetry),
+            // Throw error after exceed number of attempts
+            concatMap(o => {
+              if (this.increment > this.maxRetry) {
+                this.ngProgress.done();
+                this.alertService.closeMessage();
+                // tslint:disable-next-line: max-line-length
+                return throwError(`Failed to get expected results in ${this.maxRetry} attempts, tried with incremental time delay starting with 2s, for installing plugin ${pluginName}`);
+              }
+              return of(o);
+            }),
+          ))
+      ).subscribe(() => {
+        this.ngProgress.done();
+        this.alertService.closeMessage();
+        this.alertService.success(`Plugin ${pluginName} installed successfully.`);
+        this.getInstalledFilterPlugins(true);
+        this.serviceForm.controls.pluginToInstall.enable();
+        this.serviceForm.controls.plugin.enable();
+        this.disabledBtn = false;
+      });
   }
 
   getDescription(selectedPlugin) {
@@ -455,5 +511,18 @@ export class AddFilterWizardComponent implements OnInit {
     } else {
       this.serviceForm.controls.pluginToInstall.reset();
     }
+  }
+
+  /**
+   * Open readthedocs.io documentation of filter plugins
+   * @param selectedPlugin Selected filter plugin 
+   * 
+   */
+  goToLink(selectedPlugin: string) {
+    const pluginInfo = {
+      name: selectedPlugin,
+      type: 'filter'
+    };
+    this.docService.goToPluginLink(pluginInfo);
   }
 }