+++ src/app/components/core/north/add-task-wizard/add-task-wizard.component.ts
@@ -4,11 +4,13 @@ import { Router } from '@angular/router';
 import { assign, cloneDeep, reduce, sortBy, map } from 'lodash';
 import { Subscription } from 'rxjs';
 
-import { AlertService, SchedulesService, SharedService, PluginService, ProgressBarService } from '../../../../services';
+import { AlertService, SchedulesService, SharedService, PluginService, ProgressBarService,
+  ServicesApiService} from '../../../../services';
 import Utils from '../../../../utils';
 import { ViewConfigItemComponent } from '../../configuration-manager/view-config-item/view-config-item.component';
 import { ViewLogsComponent } from '../../packages-log/view-logs/view-logs.component';
 import { ValidateFormService } from '../../../../services/validate-form.service';
+import { DocService } from '../../../../services/doc.service';
 
 @Component({
   selector: 'app-add-task-wizard',
@@ -32,6 +34,7 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
   public selectedPluginDescription = '';
   public plugin: any;
   public showSpinner = false;
+  public isService = false;
   private subscription: Subscription;
 
   public taskType = 'North';
@@ -60,7 +63,9 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
     private router: Router,
     private ngProgress: ProgressBarService,
     private validateFormService: ValidateFormService,
-    private sharedService: SharedService
+    private sharedService: SharedService,
+    private servicesApiService: ServicesApiService,
+    private docService: DocService
   ) { }
 
   ngOnInit() {
@@ -179,7 +184,7 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
           return formValues['name'].trim() === item.name;
         });
         if (isTaskNameExist) {
-          this.alertService.error('A north task instance or south service already exists with this name.');
+          this.alertService.error('A service/task already exists with this name.');
           return false;
         }
 
@@ -205,7 +210,15 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
         previousButton.textContent = 'Previous';
         break;
       case 3:
-        this.addScheduledTask(this.payload);
+        if (this.isService) {
+          delete this.payload.schedule_repeat;
+          delete this.payload.schedule_type;
+          delete this.payload.schedule_enabled;
+          this.payload.enabled = this.isScheduleEnabled;
+          this.addService(this.payload);
+        } else {
+          this.addScheduledTask(this.payload);
+        }
         break;
       default:
         break;
@@ -314,6 +327,29 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
         });
   }
 
+  public addService(payload) {
+    this.taskForm.get('name').markAsTouched();
+    /** request started */
+    this.ngProgress.start();
+    this.servicesApiService.addService(payload)
+      .subscribe(
+        (response) => {
+          /** request done */
+          this.ngProgress.done();
+          this.alertService.success(response['name'] + ' service added successfully.', true);
+          this.router.navigate(['/north']);
+        },
+        (error) => {
+          /** request done */
+          this.ngProgress.done();
+          if (error.status === 0) {
+            console.log('service down ', error);
+          } else {
+            this.alertService.error(error.statusText);
+          }
+        });
+  }
+
   /**
    * Get edited configuration from view config child page
    * @param changedConfig changed configuration of a selected plugin
@@ -390,12 +426,12 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
   }
 
   onCheckboxClicked(event) {
-    if (event.target.checked) {
-      this.isScheduleEnabled = true;
-    } else {
-      this.isScheduleEnabled = false;
-    }
+    this.isScheduleEnabled = event.target.checked ? true : false;
     this.payload.schedule_enabled = this.isScheduleEnabled;
+  }
+
+  onServiceCheckboxClicked(event) {
+    this.isService = event.target.checked ? true : false;
   }
 
   public getSchedules(): void {
@@ -443,12 +479,14 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
 
   selectInstalledPlugin() {
     const select = <HTMLSelectElement>document.getElementById('pluginSelect');
-    for (let i = 0, j = select.options.length; i < j; ++i) {
-      if (select.options[i].innerText.toLowerCase() === this.pluginData.pluginName.toLowerCase()) {
-        this.taskForm.controls['plugin'].setValue([this.plugins[i].name]);
-        select.selectedIndex = i;
-        select.dispatchEvent(new Event('change'));
-        break;
+    if (select) {
+      for (let i = 0, j = select.options.length; i < j; ++i) {
+        if (select.options[i].innerText.toLowerCase() === this.pluginData.pluginName.toLowerCase()) {
+          this.taskForm.controls['plugin'].setValue([this.plugins[i].name]);
+          select.selectedIndex = i;
+          select.dispatchEvent(new Event('change'));
+          break;
+        }
       }
     }
   }
@@ -459,6 +497,14 @@ export class AddTaskWizardComponent implements OnInit, OnDestroy {
 
   public hideLoadingSpinner() {
     this.showSpinner = false;
+  }
+
+  goToLink() {
+    const pluginInfo = {
+      name: this.plugin,
+      type: 'North'
+    };
+    this.docService.goToPluginLink(pluginInfo);
   }
 
   ngOnDestroy() {