+++ src/app/components/core/notifications/notifications.component.ts
@@ -33,6 +33,7 @@ export class NotificationsComponent implements OnInit, OnDestroy {
   public notificationServiceRecord: any;
   public availableServices = [];
   private subscription: Subscription;
+  private modalSub: Subscription;
   private viewPortSubscription: Subscription;
   public showSpinner = false;
   public notificationServiceData = {};
@@ -54,6 +55,7 @@ export class NotificationsComponent implements OnInit, OnDestroy {
     private sharedService: SharedService) { }
 
   ngOnInit() {
+    this.onNotifySettingModal();
     this.checkNotificationServiceStatus();
     this.getNotificationInstance();
     this.subscription = this.sharedService.showLogs.subscribe(showPackageLogs => {
@@ -129,18 +131,22 @@ export class NotificationsComponent implements OnInit, OnDestroy {
     }, 2000);
   }
 
-  onNotifySettingModal(event: any) {
-    if (event.isEnabled !== undefined) {
-      this.isNotificationServiceEnabled = event.isEnabled;
-    }
-    if (event.isAddDeleteAction !== undefined) {
-      setTimeout(() => {
-        this.checkNotificationServiceStatus(true);
-      }, 2000);
-    }
-    if (event.isConfigChanged !== undefined) {
-      this.checkServiceStatus();
-    }
+  onNotifySettingModal() {
+    this.modalSub = this.notificationService.notifyServiceEmitter
+      .subscribe(event => {
+        if (event === null) { return; }
+        if (event.isEnabled !== undefined) {
+          this.isNotificationServiceEnabled = event.isEnabled;
+        }
+        if (event.isAddDeleteAction !== undefined) {
+          setTimeout(() => {
+            this.checkNotificationServiceStatus(true);
+          }, 2000);
+        }
+        if (event.isConfigChanged !== undefined) {
+          this.checkServiceStatus();
+        }
+      });
   }
 
   public showLoadingSpinner() {
@@ -265,11 +271,14 @@ export class NotificationsComponent implements OnInit, OnDestroy {
     this.notificationServiceModal.toggleModal(true);
   }
 
-  goToLink() {
-    this.docService.goToNotificationDocLink();
+  goToLink(urlSlug: string) {
+    this.docService.goToNotificationDocLink(urlSlug);
   }
 
   ngOnDestroy() {
+    if (this.modalSub) {
+      this.modalSub.unsubscribe();
+    }
     this.subscription.unsubscribe();
     this.viewPortSubscription.unsubscribe();
   }