+++ src/app/components/core/system-log/system-log.component.ts
@@ -12,7 +12,7 @@ import { POLLING_INTERVAL } from '../../../utils';
 export class SystemLogComponent implements OnInit, OnDestroy {
   public logs: any;
   public source = '';
-  public level = '';
+  public level = 'info';
   public totalCount: any;
   DEFAULT_LIMIT = 50;
   limit = this.DEFAULT_LIMIT;
@@ -33,16 +33,16 @@ export class SystemLogComponent implements OnInit, OnDestroy {
     private alertService: AlertService,
     public ngProgress: ProgressBarService,
     private ping: PingService) {
-      this.isAlive = true;
-      this.ping.pingIntervalChanged
-        .pipe(takeUntil(this.destroy$))
-        .subscribe((timeInterval: number) => {
-          if (timeInterval === -1) {
-            this.isAlive = false;
-          }
-          this.refreshInterval = timeInterval;
-        });
-    }
+    this.isAlive = true;
+    this.ping.pingIntervalChanged
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((timeInterval: number) => {
+        if (timeInterval === -1) {
+          this.isAlive = false;
+        }
+        this.refreshInterval = timeInterval;
+      });
+  }
 
   ngOnInit() {
     this.getSysLogs();
@@ -63,22 +63,36 @@ export class SystemLogComponent implements OnInit, OnDestroy {
           const south_c = [];
           const notification_c = [];
           const north_c = [];
+          // processName north_C represents Northbound services
+          const north_C = [];
           const north = [];
+          const management = [];
+          const dispatcher_c = [];
+
           data.schedules.forEach(sch => {
-            if ('south_c'.includes(sch.processName)) {
+            if ('south_c' === sch.processName) {
               south_c.push(sch);
             }
-            if ('notification_c'.includes(sch.processName)) {
+            if ('notification_c' === sch.processName) {
               notification_c.push(sch);
             }
-            if ('north_c'.includes(sch.processName)) {
+            if ('north_c' === sch.processName) {
               north_c.push(sch);
             }
-            if ('north'.includes(sch.processName)) {
+            if ('north_C' === sch.processName) {
+              north_C.push(sch);
+            }
+            if ('north' === sch.processName) {
               north.push(sch);
             }
+            if ('management' === sch.processName) {
+              management.push(sch);
+            }
+            if ('dispatcher_c' === sch.processName) {
+              dispatcher_c.push(sch);
+            }
           });
-          this.scheduleData = south_c.concat(notification_c, north_c, north);
+          this.scheduleData = south_c.concat(notification_c, north_c, north_C, north, management, dispatcher_c);
         },
         error => {
           if (error.status === 0) {
@@ -194,14 +208,14 @@ export class SystemLogComponent implements OnInit, OnDestroy {
     if (filter === 'source') {
       this.source = value.trim().toLowerCase() === 'all' ? '' : value.trim();
     } else {
-      this.level = value.trim().toLowerCase() === 'info' ? '' : value.trim().toLowerCase();
+      this.level = value.trim().toLowerCase() === 'debug' ? '' : value.trim().toLowerCase();
     }
     this.getSysLogs();
   }
 
   capitalizeInitialWord(s: string) {
     return s.charAt(0).toUpperCase() + s.slice(1);
-}
+  }
 
   public getSysLogs(autoRefresh = false) {
     if (autoRefresh === false) {
@@ -218,11 +232,12 @@ export class SystemLogComponent implements OnInit, OnDestroy {
           }
           const logs = [];
           data['logs'].forEach(l => {
-            let fl = l.replace('INFO:', '<span class="tag is-light tag-syslog">INFO:</span>'); // is-info
-            fl = fl.replace('WARNING:', '<span class="tag is-warning tag-syslog">WARNING:</span>');
-            fl = fl.replace('ERROR:', '<span class="tag is-danger tag-syslog">ERROR:</span>');
-            fl = fl.replace('FATAL:', '<span class="tag is-danger tag-syslog">FATAL:</span>');
-            fl = fl.replace('EXCEPTION:', '<span class="tag is-danger tag-syslog">EXCEPTION:</span>');
+            let fl = l.replace('DEBUG:', '<span class="tag is-light tag-syslog">DEBUG:</span>');
+            fl = l.replace('INFO:', '<span class="tag is-white tag-syslog">INFO:</span>'); // is-info
+            fl = fl.replace('WARNING:', '<span class="tag is-light is-warning tag-syslog">WARNING:</span>');
+            fl = fl.replace('ERROR:', '<span class="tag is-light is-danger tag-syslog">ERROR:</span>');
+            fl = fl.replace('FATAL:', '<span class="tag is-light is-danger tag-syslog">FATAL:</span>');
+            fl = fl.replace('EXCEPTION:', '<span class="tag is-light is-danger tag-syslog">EXCEPTION:</span>');
             logs.push(fl);
           });
           this.logs = logs.reverse();