+++ src/app/components/layout/footer/footer.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 
 import { environment } from '../../../../environments/environment';
 import * as data from '../../../../git-version.json';
+import * as moment from 'moment';
 
 @Component({
   selector: 'app-footer',
@@ -13,6 +14,7 @@ import * as data from '../../../../git-version.json';
 export class FooterComponent {
   public appVersion: string = environment.VERSION;
   public git = data['default'];
+  copyrightYear = moment().year();
 
   constructor(public router: Router) {}
 }