+++ src/app/services/shared.service.ts
@@ -10,4 +10,5 @@ export class SharedService {
   public theme: BehaviorSubject<any> = new BehaviorSubject<any>(localStorage.getItem('OPTED_THEME') != null ?
   localStorage.getItem('OPTED_THEME') : 'light');
   public viewport: BehaviorSubject<any> = new BehaviorSubject<any>(false);
+  public connectionInfo: BehaviorSubject<any> = new BehaviorSubject<any>(false);
 }